Add more strict warning flags

+++ blieng/bliany.h
@@ -69,7 +69,7 @@ public:
         if (this->type() == typeid(int)) return boost::any_cast<int>(*this);
         if (this->type() == typeid(unsigned int)) return boost::any_cast<unsigned int>(*this);
         if (this->type() == typeid(long)) return boost::any_cast<long>(*this);
-        if (this->type() == typeid(unsigned long)) return boost::any_cast<unsigned long>(*this);
+        if (this->type() == typeid(unsigned long)) return static_cast<long>(boost::any_cast<unsigned long>(*this));
         if (this->type() == typeid(float)) return boost::any_cast<float>(*this);
         if (this->type() == typeid(double)) return boost::any_cast<double>(*this);
         //std::cerr << "Error, invalid value, can't convert to a number\n";
